完成部分班组功能

+++ EquipManage.Application/SystemDocument/OperationClassApp.cs
@@ -7,6 +7,7 @@
 using EquipManage.Domain.Entity.SystemDocument;
 using EquipManage.Domain.IRepository.SystemDocument;
 using EquipManage.Repository.SystemDocument;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -15,6 +16,7 @@ namespace EquipManage.Application.SystemDocument
     public class OperationClassApp
     {
         private IOperationClassRepository service = new OperationClassRepository();
+        private IOperationClassMemberRepository detailService = new OperationClassMemberRepository();
         public List<OperationClassEntity> GetList()
         {
             return service.IQueryable().OrderBy(t => t.FCreatorTime).ToList();
@@ -38,7 +40,14 @@ namespace EquipManage.Application.SystemDocument
         }
         public void DeleteForm(string keyValue)
         {
+            if (detailService.IQueryable().Count(t => t.FOperationClassID.Equals(keyValue)) > 0)
+            {
+                throw new Exception("删除失败！操作的对象包含了班组成员。");
+            }
+            else
+            {
                 service.Delete(t => t.FId == keyValue);
+            }
         }
         public void UpdateForm(OperationClassEntity operationClassEntity)
         {
